requirement.txt
tqdm==4.65.0
neurokit2==0.2.6
matplotlib==3.7.2
pytorch==2.0.1
numpy==1.26.0
panda==1.5.3